requirements.txt
# Core requirements
pyqt6==6.6.1
pyqt6-charts==6.6.0
python-nmap==0.7.1
pyattck==6.1.1
requests==2.31.0
openai==1.12.0
scapy==2.5.0
python-whois==0.8.0
netifaces==0.11.0
pymetasploit3==1.0.3

# For documentation and development
sphinx==7.2.6
black==24.1.1
pytest==7.4.3